fixup! [FLINK-16326] [core] Validate strictly required Flink configs in StatefulFunctionsConfig
